git clone args

+++ shell/src/bin/git_clone.rs
@@ -1,6 +1,27 @@
+#![allow(dead_code, unused_imports, non_camel_case_types)]
+
+const USAGE: &str = "
+Usage: quicksort [eagle] [options]
+       quicksort --help
+
+clone tag
+
+Commands:
+    eagle            Run the benchmark in different modes and print the timings.
+
+Options:
+    --thread-size N     Number of 32-bit words to sort [default: 4] (1GB)
+    -h, --help          Show this message.
+    --source, -s N          source json path
+    --dir, -d N            root dir which all repositories cloned in [default: d:/eagle_repos/]
+    --result, -r N          json path which result written into
+    --all-tags, -a      clone all tags
+    --clone-new, -c     remove old project and clone new one
+";
+
 use std::{env, fs, thread};
 use std::ops::Index;
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 use std::process::Command;
 use std::str::FromStr;
 use serde_json::{json, Value};
@@ -8,9 +29,31 @@ use std::sync::{Arc, mpsc, Mutex};
 use std::sync::mpsc::Sender;
 use std::time::Duration;
 
+#[derive(serde::Deserialize, Debug)]
+pub struct Args {
+    cmd_eagle: bool,
+    flag_source: String,
+    flag_result: String,
+    flag_dir: String,
+    flag_thread_size: usize,
+    flag_all_tags: bool,
+    flag_clone_new: bool
+}
+
+use docopt::Docopt;
+use std::time::Instant;
+
 fn main() {
-    let path = env::args().nth(1).expect("no path");
-    let s = fs::read_to_string(path).expect("Failed to read");
+    let args = env::args().collect::<Vec<String>>();
+    let args: Args = Docopt::new(USAGE)
+        .and_then(|d| d.argv(args).deserialize())
+        .unwrap_or_else(|e| e.exit());
+    println!("args = {:?}", args);
+    clone_all(args);
+}
+
+fn clone_all(args: Args) {
+    let s = fs::read_to_string(&args.flag_source).expect("Failed to read");
     let arr_value: Value = serde_json::from_str(&s).expect("not value");
     // println!("{:?}", arr_value);
     let (ts, tr) = mpsc::channel();
@@ -18,10 +61,12 @@ fn main() {
     let counter = Arc::new(Mutex::new(0));
     let mut handles = vec![];
     let arr = Arc::new(arr_value.as_array().unwrap().clone());
-    for _ in 0..4 {
+    let args = Arc::new(args);
+    for _ in 0..args.flag_thread_size {
         let counter = Arc::clone(&counter);
         let ts = mpsc::Sender::clone(&ts);
         let arr = Arc::clone(&arr);
+        let args= Arc::clone(&args);
         let handle = thread::spawn(move || {
             loop {
                 let index;
@@ -36,7 +81,7 @@ fn main() {
                 }
                 let value = &(*arr);
                 let clone_obj = &value[index as usize];
-                clone_one(&ts, clone_obj);
+                clone_one(&ts, clone_obj, &args);
             }
         });
         handles.push(handle);
@@ -51,46 +96,77 @@ fn main() {
         handle.join().unwrap();
     }
 
-    fs::write("tmp.json", res.to_string()).expect("Failed to write");
+    fs::write(&args.flag_result, res.to_string()).expect("Failed to write");
 }
 
-fn clone_one(sender: &Sender<Value>, value: &Value) {
+fn clone_one(sender: &Sender<Value>, value: &Value, args: &Arc<Args>) {
     let obj = value.as_object().expect("not obj");
     let url = obj.get("url").expect("no url").as_str().expect("not str");
     let tag = obj.get("tag").expect("no tag").as_str().expect("not str");
-    let path = url.replace("https://", "d:/tmp/");
+    let path = url.replace("https://", &args.flag_dir);
     let path_buf = PathBuf::from_str(&path).expect("not path");
+    let mut js = value.clone();
     if path_buf.exists() {
-        fs::remove_dir_all(path_buf).unwrap();
+        if args.flag_clone_new {
+            fs::remove_dir_all(path_buf).unwrap();
+        } else {
+            js["clone_ok"] = json!(0);
+            js["clone_msg"] = json!("ok");
+            js["project_path"] = json!(&path);
+            sender.send(js);
+            return;
+        }
     }
+
     let mut c = Command::new("git");
     c.args(["clone", "--depth", "1", "--branch", tag, url, &path]);
     // --depth 1 --branch <tag_name> <repo_url>
     let output = c.output().expect("no output");
     let code = output.status.code().unwrap_or(-1);
-    let mut clone_ok = true;
+
     if code != 0 {
-        clone_ok = false;
-    }
-    let mut js = value.clone();
-    js["clone_ok"] = json!(clone_ok);
-    js["clone_msg"] = json!("ok");
-    if !clone_ok {
-        let msg;
+        js["clone_ok"] = json!(code);
         unsafe {
-            msg = String::from_utf8_unchecked(output.stderr.clone());
+            let msg = String::from_utf8_unchecked(output.stderr.clone());
+            js["clone_msg"] = json!(&msg);
         }
-        js["clone_msg"] = json!(&msg);
+    } else {
+        js["clone_ok"] = json!(0);
+        js["clone_msg"] = json!("ok");
+        js["project_path"] = json!(&path);
     }
-    println!("{:?}", output);
+
+    println!("output.status = {:?}", output.status);
     sender.send(js);
+}
+
+pub fn git_cur_tag(path: &str) -> Option<String> {
+    let mut c = Command::new("git");
+    c.args(["-C", path, "branch"]);
+    let output = c.output();
+    if let Ok(o) = output {
+        let msg = String::from_utf8(o.stdout).unwrap();
+        return Some(msg);
+    }
+    return None;
 }
 
 #[cfg(test)]
 mod tests {
+    use crate::git_cur_tag;
+
     #[test]
     fn test_a() {
         let a = vec!["aaa"];
         let b = a[0];
+    }
+
+    #[test]
+    fn test_git_cur_tag() {
+        let p = "d:/tmp/gitee.com/egzosn/pay-java-parent";
+        let a = git_cur_tag(p);
+        if let Some(a) = a {
+            println!("{}", a);
+        }
     }
 }
